# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Social Microservice Api
  description: Description
  version: 1.0.0

tags:
  - name: chat
  - name: feed
  - name: posts
  - name: comments
  - name: reaction
  - name: connections

servers:
  - url: 'localhost:8080/service/social'
    description: Production server (uses live data)

paths:
  '/feed/{userId}':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
    get:
      summary: Get the all posts of the specified user
      tags:
        - feed
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedDto'
      operationId: get-feed
  /feed:
    parameters: [ ]
    get:
      summary: Get all posts of the current user's friends and followers
      tags:
        - feed
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedDto'
      description: ''
  /friends:
    parameters: [ ]
    get:
      summary: Get current user's friends
      tags:
        - connections
      responses:
        '200':
          description: Returns the names of the current user's friends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    post:
      summary: Add a friend to the current user
      tags:
        - connections
      responses:
        '200':
          description: Desc

  /posts:
    post:
        summary: 'Create a new post from the current user'
        tags:
          - posts
        operationId: post-post
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPostDto'
          description: ''
        responses:
          '201':
            description: A new post has been created
  /posts/{postId}:
    parameters:
      - schema:
          type: string
          format: uuid
        name: postId
        in: path
        required: true
        description: id of the post
    patch:
      summary: "Update a post's content"
      tags:
        - posts
      operationId: patch-post
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPatchDto'
        description: ''
      description: ''
    delete:
      summary: 'Deletes specified post'
      tags:
        - posts
      operationId: delete-post
      responses:
        '200':
          description: OK
  /comments:
    post:
      summary: ''
      tags:
        - comments
      operationId: post-comments
      responses:
        '201':
          description: Created
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentPostDto'
  '/comments/{commentId}':
    parameters:
      - schema:
          type: string
        name: commentId
        in: path
        required: true
    patch:
      summary: 'Update the content of a post'
      tags:
        - comments
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentPatchDto'
    delete:
      summary: ''
      tags:
        - comments
      operationId: delete-comments-commentId
      responses:
        '200':
          description: OK
  /reactions:
    put:
      summary: 'Adds a reaction of the given type to the specified item'
      tags:
        - reactions
      operationId: put-reactions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionDto'
        description: ''
  '/reactions/{reactionId}':
    parameters:
      - schema:
          type: string
        name: reactionId
        in: path
        required: true
    delete:
      summary: ''
      tags:
        - reactions
      operationId: delete-reactions-reactionId
      responses:
        '200':
          description: OK

components:
  schemas:
    FeedDto:
      title: FeedDto
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostResponseDto'
    PostPostDto:
      title: PostRequestDto
      type: object
      properties:
        createdBy:
          type: string
          format: uuid
        content:
          type: string
        contentType:
          $ref: '#/components/schemas/PostContentTypeDto'
        isPrivate:
          type: boolean
        taggedUsers:
          type: array
          default: []
          items:
            type: string
            format: uuid
      required:
        - createdBy
        - content
        - contentType
        - isPrivate
    PostPatchDto:
      type: object
      title: PostPatchDto
      properties:
        content:
          type: string
        contentType:
          $ref: '#/components/schemas/PostContentTypeDto'
        isPrivate:
          type: boolean
    PostResponseDto:
      type: object
      title: PostResponseDto
      properties:
        postId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        content:
          type: string
        timestamp: { }
        isPrivate:
          type: boolean
        taggedUserIds:
          type: array
          items:
            type: string
            format: uuid
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponseDto'
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/ReactionDto'
    CommentPostDto:
      title: CommentPostDto
      type: object
      properties:
        parentComment:
          type: string
        content:
          type: string
        parent:
          type: string
        parentType:
          $ref: '#/components/schemas/ParentType'
        contentType:
          $ref: '#/components/schemas/CommentContentType'
      required:
        - parent
        - content
        - parentType
        - contentType
    CommentPatchDto:
      title: CommentPatchDto
      type: object
      properties:
        content:
          type: string
        contentType:
          $ref: '#/components/schemas/CommentContentType'
      required:
        - content
        - contentType
    CommentResponseDto:
      title: CommentResponseDto
      type: object
      properties:
        commentId:
          type: string
        userId:
          type: string
        content:
          type: string
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/ReactionDto'
    ReactionDto:
      title: ReactionDto
      type: object
      properties:
        userId:
          type: string
        parentId:
          type: string
        parentType:
          type: string
        reactionType:
          type: string
          enum:
            - like
            - love
            - haha
            - wow
            - sad
            - angry
    UserDto:
      title: User
      type: object
      properties:
        userId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        userType:
          type: string
          enum:
            - follower
            - friend
    UsersDto:
      title: UsersDto
      type: object
      properties:
        users:
          type: object
          items:
            $ref: '#/components/schemas/UserDto'
    ParentType:
      title: ParentType
      type: string
      enum:
        - post
        - news
    CommentContentType:
      title: ContentType
      type: string
      enum:
        - text
        - gif
        - image
    PostContentTypeDto:
      title: PostType
      type: string
      enum:
        - text
        - gif
        - image
        - video